requirements.txt
Flask==3.0.0
scikit-learn==1.3.2
pandas==2.1.4
joblib==1.3.2
scikit-fuzzy
numpy
tensorflow-cpu
